+++ algorithms/0151-reverse-words-in-a-string.go
@@ -1,25 +1,11 @@
 package algorithms
 
-import (
-	"bytes"
-)
+import "strings"
 
 func reverseWords(s string) string {
-	var buff bytes.Buffer
-	for h, t := len(s), len(s); t > 0; t = h {
-		// find word end
-		for ; t > 0 && s[t-1] == ' '; t-- {
-		}
-		// find word begin
-		for h = t; h > 0 && s[h-1] != ' '; h-- {
-		}
-		// store word, if exists
-		if h < t {
-			if buff.Len() > 0 {
-				buff.WriteByte(' ')
-			}
-			buff.WriteString(s[h:t])
-		}
+	fileds := strings.Fields(s)
+	for i, j := 0, len(fileds)-1; i < j; i, j = i+1, j-1 {
+		fileds[i], fileds[j] = fileds[j], fileds[i]
 	}
-	return buff.String()
+	return strings.Join(fileds, " ")
 }